filter.io: Move functions related to I/O here
