+++ augur/filter/_run.py
@@ -21,6 +21,7 @@ from augur.io.metadata import read_metadata
 from augur.io.sequences import read_sequences, write_sequences
 from augur.io.vcf import is_vcf as filename_is_vcf, write_vcf
 from augur.utils import read_strains
+from .io import cleanup_outputs, filter_kwargs_to_str, read_priority_scores
 
 from . import GROUP_BY_GENERATED_COLUMNS
 
@@ -29,21 +30,6 @@ SEQUENCE_ONLY_FILTERS = (
     "min_length",
     "non_nucleotide",
 )
-
-
-def read_priority_scores(fname):
-    def constant_factory(value):
-        return lambda: value
-
-    try:
-        with open(fname, encoding='utf-8') as pfile:
-            return defaultdict(constant_factory(-np.inf), {
-                elems[0]: float(elems[1])
-                for elems in (line.strip().split('\t') if '\t' in line else line.strip().split() for line in pfile.readlines())
-            })
-    except Exception as e:
-        print(f"ERROR: missing or malformed priority scores file {fname}", file=sys.stderr)
-        raise e
 
 # Define metadata filters.
 
@@ -648,56 +634,6 @@ def construct_filters(args, sequence_index):
         ))
 
     return exclude_by, include_by
-
-
-def filter_kwargs_to_str(kwargs):
-    """Convert a dictionary of kwargs to a JSON string for downstream reporting.
-
-    This structured string can be converted back into a Python data structure
-    later for more sophisticated reporting by specific kwargs.
-
-    This function excludes data types from arguments like pandas DataFrames and
-    also converts floating point numbers to a fixed precision for better
-    readability and reproducibility.
-
-    Parameters
-    ----------
-    kwargs : dict
-        Dictionary of kwargs passed to a given filter function.
-
-    Returns
-    -------
-    str :
-        String representation of the kwargs for reporting.
-
-
-    >>> sequence_index = pd.DataFrame([{"strain": "strain1", "ACGT": 28000}, {"strain": "strain2", "ACGT": 26000}, {"strain": "strain3", "ACGT": 5000}]).set_index("strain")
-    >>> exclude_by = [(filter_by_sequence_length, {"sequence_index": sequence_index, "min_length": 27000})]
-    >>> filter_kwargs_to_str(exclude_by[0][1])
-    '[["min_length", 27000]]'
-    >>> exclude_by = [(filter_by_date, {"max_date": numeric_date("2020-04-01"), "min_date": numeric_date("2020-03-01")})]
-    >>> filter_kwargs_to_str(exclude_by[0][1])
-    '[["max_date", 2020.25], ["min_date", 2020.17]]'
-
-    """
-    # Sort keys prior to processing to guarantee the same output order
-    # regardless of the input order.
-    sorted_keys = sorted(kwargs.keys())
-
-    kwarg_list = []
-    for key in sorted_keys:
-        value = kwargs[key]
-
-        # Handle special cases for data types that we want to represent
-        # differently from their defaults or not at all.
-        if isinstance(value, pd.DataFrame):
-            continue
-        elif isinstance(value, float):
-            value = round(value, 2)
-
-        kwarg_list.append((key, value))
-
-    return json.dumps(kwarg_list)
 
 
 def apply_filters(metadata, exclude_by, include_by):
@@ -1412,7 +1348,7 @@ def run(args):
             set(metadata.index[metadata.index.isin(metadata_strains)])
         )
         if len(duplicate_strains) > 0:
-            _cleanup_outputs(args)
+            cleanup_outputs(args)
             raise AugurError(f"The following strains are duplicated in '{args.metadata}':\n" + "\n".join(sorted(duplicate_strains)))
 
         # Maintain list of all strains seen.
@@ -1909,23 +1845,3 @@ def _calculate_fractional_sequences_per_group(
             hi = mid
 
     return (lo + hi) / 2
-
-
-def _cleanup_outputs(args):
-    """Remove output files. Useful when terminating midway through a loop of metadata chunks."""
-    if args.output:
-        _try_remove(args.output)
-    if args.output_metadata:
-        _try_remove(args.output_metadata)
-    if args.output_strains:
-        _try_remove(args.output_strains)
-    if args.output_log:
-        _try_remove(args.output_log)
-
-
-def _try_remove(filepath):
-    """Remove a file if it exists."""
-    try:
-        os.remove(filepath)
-    except FileNotFoundError:
-        pass
